changed the Rating property back to an int as it uses stars and fixed the implementation in Saved Recipe as it did not trigger a save when only the Rating was changed.

+++ RecipeApp/Interfaces/IRecipe.cs
@@ -5,6 +5,6 @@ public interface IRecipe
     string Title { get; }
     string? Category { get; }
     string? ImageUrl { get; }
-    double Rating { get; }
+    int Rating { get; }
     string? Description { get; }
 }
